+++ src/main/java/pl/edu/agh/to/cinemawiet/screening/model/ScreeningRequest.java
@@ -2,5 +2,5 @@ package pl.edu.agh.to.cinemawiet.screening.model;
 
 import java.sql.Timestamp;
 
-public record ScreeningRequest(int filmId, int hallId, Timestamp screeningDate) {
+public record ScreeningRequest(long filmId, long hallId, Timestamp screeningDate) {
 }
